requirements.txt
flask==2.2.3
flask-socketio==5.3.2
python-dotenv==1.0.0
gunicorn==20.1.0
eventlet==0.33.3
SpeechRecognition==3.10.0
pydub==0.25.1
nltk==3.8.1
numpy==1.24.2
networkx==3.0
pandas==1.5.3
scikit-learn==1.2.2
python-engineio==4.3.4
python-socketio==5.7.2